update for two clocks, fix wire chamber reco

+++ interface/TOFPETRecoTree.h
@@ -37,8 +37,8 @@ public:
     int yi[64];
     int spillNum;
     double t_h4daq;
-    double t_totrigger[64];
-    double t_tofpet[64];
+    long long int t_totrigger[64];
+    long long int t_tofpet[64];
     double deltaT_h4daq_trigger;
 };
